refactor(filter): 修改分页参数类型从 int64 为 int
将 Filter 接口及其实现中的 Pagination 方法的 pageNum 和 pageSize 参数类型由 int64 调整为 int，
同时更新了相关工具类 Pagination 结构体字段类型以及测试代码中对应的调用逻辑。
此变更统一了分页参数的类型使用，提升代码一致性与可维护性。

+++ models/filter.go
@@ -44,7 +44,7 @@ type Filter interface {
 	In(key string, val any) *cd.Error
 	NotIn(key string, val any) *cd.Error
 	Like(key string, val any) *cd.Error
-	Pagination(pageNum, pageSize int64)
+	Pagination(pageNum, pageSize int)
 	Sort(fieldName string, ascFlag bool)
 	ValueMask(val any) *cd.Error
